Safari-609.1.11.1/LayoutTests/fast/dom/timer-throttling-subframe-expected.txt

Tests that timers are throttled in subframes that are outside the viewport

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Frame is initially outside the viewport so timers should be throttled
PASS testFrame.contentWindow.internals.areTimersThrottled() became true
PASS internals.areTimersThrottled() is false
PASS testFrame.contentWindow.internals.areTimersThrottled() is true
PASS grandChildFrame.contentWindow.internals.areTimersThrottled() is true
Scrolling frame into view.
Timers should no longer be throttled
PASS internals.areTimersThrottled() is false
PASS testFrame.contentWindow.internals.areTimersThrottled() is false
PASS grandChildFrame.contentWindow.internals.areTimersThrottled() is false
Scrolling frame out of view again.
PASS internals.areTimersThrottled() is false
PASS testFrame.contentWindow.internals.areTimersThrottled() became true
PASS testFrame.contentWindow.internals.areTimersThrottled() is true
PASS grandChildFrame.contentWindow.internals.areTimersThrottled() is true
PASS successfullyParsed is true

TEST COMPLETE
